fix: move getting of user matches to user
