+++ app/backend/src/controllers/user.controller.js
@@ -134,3 +134,29 @@ export async function deleteUser(request, reply) {
 		return reply.code(500).send({ error: "Failed to delete user" });
 	}
 }
+
+export async function getUserMatches(request, reply) {
+	const playerId = parseInt(request.params.id, 10);
+
+	try {
+		const matches = await prisma.match.findMany({
+			where: {
+				playerId
+			},
+			select: {
+				playerNickname: true,
+				opponentNickname: true,
+				playerScore: true,
+				opponentScore: true,
+				date: true,
+			}
+		})
+		if (!matches) {
+			return reply.code(404).send({ error: "No matches found for this user" });
+		}
+		return reply.code(200).send(matches);
+	} catch (err) {
+		request.log.error(err);
+		return reply.code(500).send({ error: "Failed to retrieve matches" });
+	}
+}
